issue #91 test_common_code_abstraction
